Added "here" package to enable a mini app to run using the functions and server code in the SPBAND dashboard folder. This enables relative filepaths from sub-folders.
Added code to group the island boards for the Terminations dataset. This group is removed from the TERMINATIONS measure but used in the GESTATION AT TERMINATION measure.

Tidied up some code. Removed functions that are not used to create dataframes. Functions to create charts etc. are now read in from the dashboard folder to remove duplication.

Created a mini app to view the data as presented in the main dashboard.